Base Commit.
This is the first build released on GitHub for the TRHT Web server.

It reads everything on the index page from a template and gets that information from arrays and variables stored in RAM. These variables are updated on launch via a MySQL database and everytime an admin makes an update to a list it is read from MySQL.

The admin page uses JWT authentication which I believe to be secure as of right now. It uses stateless AuthTokens and stateful RefreshTokens to the AuthTokens are stateless to remove querying the server time but mixed with a stateful permanent token so access can be revoked easily.