Added gallery in admin.
Now from the admin panel you can update, delete and make new slides for the gallery. This also works when you don't want to change the image for updating as that is useful. Apart from that a major bug fix where the files had a different import from their pre-git version.