Exploit and bug fixes, user experience improved.
Just finishing off this project now by fixing all known exploits and bugs.

Also improved user experience as now the user sees a popup when the server calls a fail. Now the user is redirected to login instead of home when the server logs the user out.

There may be improvements and bug / expolit fixes in the future but for now this project seems finished.