Improved JTI system.
Before the JTI system was just a map with tokens in; now it uses a MySQL DataBase. I've enabled the garbage collecter as for some reason it wasn't being run before. I've tested this new system and it works well. The main reason for implementing this is if the server had to restart for an update or the likes every refresh token was invalidated as their JTI was. This solution solves this problem: it's using the SSD not RAM.

TL;DR: Improves user experience.