Logout security improvement and bug fixes.
The logout will now invalidate the refresh token aswell. This is just a small security improvement: incase the token is stolen it'll now be killed.

Bug fix: before when doing user changes it would use the token's priv but the page would use the MySQL DB. Meaning for upto 15 minutes it could throw errors and cause inconvenience. The solution is now tokens don't contain a priv int instead the page gets the priv from getting a user based on the token's subject. The priv check is now based on this.